refactor auth logic

+++ src/main/java/com/irma/mysheila/repositories/CategoryRepository.java
@@ -3,6 +3,5 @@ package com.irma.mysheila.repositories;
 import com.irma.mysheila.entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 }
